Refacto to use services and an axios instance
