+++ PyroApp/services/auth.service.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import apiClient from './apiClient.service';
 
 async function saveToken(token: string): Promise<void> {
   await AsyncStorage.setItem('token', token);
@@ -17,9 +18,27 @@ async function isLogged(): Promise<boolean> {
   return !!token;
 }
 
+async function login() {
+  apiClient
+    .post(
+      '/login/access-token',
+      {
+        username: 'CentraleSupelec',
+        password: '&Fib7!Eoc!LSfMpa',
+      },
+      {
+        headers: {'Content-Type': 'application/x-www-form-urlencoded'},
+      },
+    )
+    .then(response => {
+      saveToken(response.data.access_token);
+    })
+    .catch(error => console.log(error));
+}
+
 export const authService = {
-  saveToken,
   getToken,
   removeToken,
   isLogged,
+  login,
 };
